Removed unused fns

+++ 494TaskScheduler/494TaskScheduler/Task.cpp
@@ -21,19 +21,11 @@ double Task::getWorstCaseExecutionTime() { return f_wcet; }
 
 double Task::getUtlization() { return (f_wcet / (double) i_period); }
 
-//double Task::getAbsoluteDeadline() { return f_absoluteDeadline; }
-
 int Task::getNextArrivalTime() {return i_nextArrival;}
 
-//int Task::getTimeKeeper() {return i_timeKeeper;}
-
-void Task::incrementProcessorTimeConsumed(float incrementValue) { f_processorTimeConsumed+=incrementValue; }
-
-//void Task::updateAbsoluteDeadline(int time) {f_absoluteDeadline = time;}
+void Task::incrementProcessorTimeConsumed(double incrementValue) { f_processorTimeConsumed+=incrementValue; }
 
 void Task::updateNextArrivalTime(int time) {i_nextArrival = time;}
-
-//void Task::timeKeep(int time) {i_timeKeeper = time;}
 
 //void Task::complete(bool complete) {b_isComplete = complete;}
